// frontend/src/gen/api/src/models/MediaByUserIdDtoGenericListNewLevelResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MediaByUserIdDtoGenericList } from './MediaByUserIdDtoGenericList';
import {
    MediaByUserIdDtoGenericListFromJSON,
    MediaByUserIdDtoGenericListFromJSONTyped,
    MediaByUserIdDtoGenericListToJSON,
} from './MediaByUserIdDtoGenericList';

/**
 * 
 * @export
 * @interface MediaByUserIdDtoGenericListNewLevelResponse
 */
export interface MediaByUserIdDtoGenericListNewLevelResponse {
    /**
     * 
     * @type {boolean}
     * @memberof MediaByUserIdDtoGenericListNewLevelResponse
     */
    isSuccess?: boolean;
    /**
     * 
     * @type {string}
     * @memberof MediaByUserIdDtoGenericListNewLevelResponse
     */
    message?: string;
    /**
     * 
     * @type {MediaByUserIdDtoGenericList}
     * @memberof MediaByUserIdDtoGenericListNewLevelResponse
     */
    data?: MediaByUserIdDtoGenericList;
}

/**
 * Check if a given object implements the MediaByUserIdDtoGenericListNewLevelResponse interface.
 */
export function instanceOfMediaByUserIdDtoGenericListNewLevelResponse(value: object): boolean {
    return true;
}

export function MediaByUserIdDtoGenericListNewLevelResponseFromJSON(json: any): MediaByUserIdDtoGenericListNewLevelResponse {
    return MediaByUserIdDtoGenericListNewLevelResponseFromJSONTyped(json, false);
}

export function MediaByUserIdDtoGenericListNewLevelResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MediaByUserIdDtoGenericListNewLevelResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'isSuccess': json['isSuccess'] == null ? undefined : json['isSuccess'],
        'message': json['message'] == null ? undefined : json['message'],
        'data': json['data'] == null ? undefined : MediaByUserIdDtoGenericListFromJSON(json['data']),
    };
}

export function MediaByUserIdDtoGenericListNewLevelResponseToJSON(value?: MediaByUserIdDtoGenericListNewLevelResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'isSuccess': value['isSuccess'],
        'message': value['message'],
        'data': MediaByUserIdDtoGenericListToJSON(value['data']),
    };
}